backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
snowflake-connector-python==3.4.0
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
fhir-parser==1.4.0
pandas==2.1.3